refactor: decoupling the infra layer
