+++ src/modules/orders/routes/order.routes.ts
@@ -1,7 +1,7 @@
 import { Router } from "express";
 import { celebrate, Joi, Segments } from "celebrate"
 import OrdersController from "../controllers/OrdersController";
-import isAuthenticated from "@shared/http/middlewares/isAuthenticated";
+import isAuthenticated from "@shared/infra/http/middlewares/isAuthenticated";
 
 const ordersRouter = Router();
 const ordersController = new OrdersController()
