+++ src/modules/users/routes/users.routes.ts
@@ -1,7 +1,7 @@
 import { Router } from "express";
 import { celebrate, Joi, Segments } from "celebrate"
 import UsersController from "../controllers/UsersController";
-import isAuthenticated from "../../../shared/http/middlewares/isAuthenticated";
+import isAuthenticated from "../../../shared/infra/http/middlewares/isAuthenticated";
 
 const usersRouter = Router();
 const usersController = new UsersController()
