+++ src/shared/infra/http/server.ts
@@ -6,8 +6,8 @@ import { errors } from 'celebrate';
 import cors from 'cors';
 import routes from './routes';
 import AppError from '@shared/errors/AppError';
-import '@shared/typeorm';
-import rateLimiter from "@shared/http/middlewares/rateLimiter";
+import '@shared/infra/typeorm';
+import rateLimiter from "@shared/infra/http/middlewares/rateLimiter";
 
 const app = express();
